fix[venom]: invalid jump error (#4214)
fix an issue where `ret` would generate a jump to an invalid location
because the stack is not clean. the solution is to always pop
instruction outputs which are not used.

note this leads to a slight performance regression (roughly 0.07%
bytecode size), since `POP`s are always generated, even in cases when the
stack does not need to be cleaned (e.g. before `STOP`).

---------

Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>
Co-authored-by: Harry Kalogirou <harkal@nlogn.eu>