feat[venom]: improve liveness computation time (#4086)
reduce time spent in liveness by only recomputing liveness for basic
blocks which need it (any of its `cfg_out` has changed).

reduces time in venom by roughly 25%

there is also a slight codesize improvement. it seems to be sensitive
to the order in which items are added to the worklist.