fix[venom]: fix invalid `phi`s after SCCP (#4181)
This commit reduces `phi` nodes which, after SCCP, refer to a
basic block which is unreachable.  This could happen when the
SCCP would replace a `jnz` by a `jmp` instruction, resulting in a
`phi` instruction which refers to a non-existent block, or a `phi`
instruction in a basic block which only has one predecessor.  This
commit reduces such `phi` instructions into `store` instructions.

---------

Co-authored-by: Harry Kalogirou <harkal@nlogn.eu>
Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>